прикрутил фильтр по сотрудникам к журналу
