+++ Source/Applications/Desktop/Vodovoz/Filters/GtkViews/CompletedDriversWarehousesEventsJournalFilterView.cs
@@ -1,19 +1,11 @@
 using System;
-using System.Linq;
 using QS.Views.GtkUI;
-using Vodovoz.Domain.Logistic.Drivers;
 using Vodovoz.ViewModels.Journals.FilterViewModels.Logistic;
 using Vodovoz.Infrastructure.Converters;
 using Gtk;
 using Vodovoz.Core.Domain.Logistics.Drivers;
-using NHibernate.Criterion;
-using NHibernate.Transform;
-using QS.Project.DB;
 using QS.Widgets.GtkUI;
-using Vodovoz.Domain.Employees;
-using Vodovoz.Infrastructure.Report.SelectableParametersFilter;
 using Vodovoz.ReportsParameters;
-using Vodovoz.ViewModels.Reports;
 
 namespace Vodovoz.Filters.GtkViews
 {
@@ -21,14 +13,12 @@ namespace Vodovoz.Filters.GtkViews
 	public partial class CompletedDriversWarehousesEventsJournalFilterView :
 		FilterViewBase<CompletedDriversWarehousesEventsJournalFilterViewModel>
 	{
-		private readonly SelectableParametersReportFilter _filter;
 		private LeftRightListView _leftRightSortListView;
 		
 		public CompletedDriversWarehousesEventsJournalFilterView(
 			CompletedDriversWarehousesEventsJournalFilterViewModel viewModel) : base(viewModel)
 		{
 			Build();
-			_filter = new SelectableParametersReportFilter(ViewModel.UoW);
 			Configure();
 		}
 
@@ -60,49 +50,13 @@ namespace Vodovoz.Filters.GtkViews
 				.InitializeFromSource();
 			dateEventRangePicker.PeriodChangedByUser += OnDateEventPeriodChangedByUser;
 
-			ConfigureSelectableFilter();
+			CreateSelectableFilter();
 			CreateAndConfigureGroupList();
 		}
 
-		private void ConfigureSelectableFilter()
+		private void CreateSelectableFilter()
 		{
-			_filter.CreateParameterSet(
-				"Сотрудники",
-				"employees",
-				new ParametersFactory(ViewModel.UoW, (filters) =>
-				{
-					SelectableEntityParameter<Employee> resultAlias = null;
-					var query = ViewModel.UoW.Session.QueryOver<Employee>();
-
-					if(filters != null && filters.Any())
-					{
-						foreach(var f in filters)
-						{
-							query.Where(f());
-						}
-					}
-
-					var employeeProjection = CustomProjections.Concat_WS(
-						" ",
-						Projections.Property<Employee>(x => x.LastName),
-						Projections.Property<Employee>(x => x.Name),
-						Projections.Property<Employee>(x => x.Patronymic)
-					);
-
-					query.SelectList(list => list
-							.Select(x => x.Id).WithAlias(() => resultAlias.EntityId)
-							.Select(employeeProjection).WithAlias(() => resultAlias.EntityTitle))
-						.OrderBy(x => x.LastName).Asc
-						.OrderBy(x => x.Name).Asc
-						.OrderBy(x => x.Patronymic).Asc
-						.TransformUsing(Transformers.AliasToBean<SelectableEntityParameter<Employee>>());
-					
-					return query.List<SelectableParameter>();
-				})
-			);
-
-			var viewModel = new SelectableParameterReportFilterViewModel(_filter);
-			var filterWidget = new SelectableParameterReportFilterView(viewModel);
+			var filterWidget = new SelectableParameterReportFilterView(ViewModel.SelectableParameterReportFilterViewModel);
 			vboxDrivers.Add(filterWidget);
 			
 			var boxFilter = (Box.BoxChild)hboxMain[vboxDrivers];
