+++ Source/Applications/Desktop/VodovozViewModels/Journals/FilterViewModels/Logistic/CompletedDriversWarehousesEventsJournalFilterViewModel.cs
@@ -1,6 +1,11 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using Autofac;
+using NHibernate.Criterion;
+using NHibernate.Transform;
 using QS.Navigation;
+using QS.Project.DB;
 using QS.Project.Filter;
 using QS.ViewModels.Control.EEVM;
 using QS.ViewModels.Dialog;
@@ -8,9 +13,10 @@ using Vodovoz.Core.Domain.Logistics.Drivers;
 using QS.ViewModels.Widgets;
 using Vodovoz.Domain.Employees;
 using Vodovoz.Domain.Logistic.Cars;
-using Vodovoz.Domain.Logistic.Drivers;
+using Vodovoz.Infrastructure.Report.SelectableParametersFilter;
 using Vodovoz.ViewModels.Factories;
 using Vodovoz.ViewModels.Journals.JournalViewModels.Logistic;
+using Vodovoz.ViewModels.Reports;
 using Vodovoz.ViewModels.ReportsParameters.Profitability;
 using Vodovoz.ViewModels.ViewModels.Logistic;
 
@@ -23,6 +29,7 @@ namespace Vodovoz.ViewModels.Journals.FilterViewModels.Logistic
 		private readonly DialogViewModelBase _journalViewModel;
 		private readonly INavigationManager _navigationManager;
 		private readonly ILifetimeScope _scope;
+		private readonly SelectableParametersReportFilter _filter;
 		private int? _completedEventId;
 		private decimal? _distanceFromScanning;
 		private DriverWarehouseEvent _driverWarehouseEvent;
@@ -44,10 +51,12 @@ namespace Vodovoz.ViewModels.Journals.FilterViewModels.Logistic
 				leftRightListViewModelFactory ?? throw new ArgumentNullException(nameof(leftRightListViewModelFactory));
 			_navigationManager = navigationManager ?? throw new ArgumentNullException(nameof(navigationManager));
 			_scope = scope ?? throw new ArgumentNullException(nameof(scope));
+			_filter = new SelectableParametersReportFilter(UoW);
 
 			ConfigureEntryViewModels();
 			SetupSorting();
-			
+			ConfigureSelectableFilter();
+
 			if(filterParameters != null)
 			{
 				SetAndRefilterAtOnce(filterParameters);
@@ -57,6 +66,7 @@ namespace Vodovoz.ViewModels.Journals.FilterViewModels.Logistic
 		public IEntityEntryViewModel DriverWarehouseEventViewModel { get; private set; }
 		public IEntityEntryViewModel CarViewModel { get; private set; }
 		public LeftRightListViewModel<GroupingNode> SortViewModel { get; private set; }
+		public SelectableParameterReportFilterViewModel SelectableParameterReportFilterViewModel { get; private set; }
 
 		public int? CompletedEventId
 		{
@@ -105,6 +115,15 @@ namespace Vodovoz.ViewModels.Journals.FilterViewModels.Logistic
 			get => _car;
 			set => UpdateFilterField(ref _car, value);
 		}
+
+		public IEnumerable<int> GetSelectedDrivers()
+		{
+			var driversSet = SelectableParameterReportFilterViewModel.ReportFilter.ParameterSets.FirstOrDefault();
+
+			return driversSet is null
+				? Array.Empty<int>()
+				: driversSet.GetIncludedParameters().Select(x => (int)x.Value).ToArray();
+		}
 		
 		private void ConfigureEntryViewModels()
 		{
@@ -127,6 +146,46 @@ namespace Vodovoz.ViewModels.Journals.FilterViewModels.Logistic
 		private void SetupSorting()
 		{
 			SortViewModel = _leftRightListViewModelFactory.CreateCompletedDriverEventsSortingConstructor();
+		}
+
+		private void ConfigureSelectableFilter()
+		{
+			_filter.CreateParameterSet(
+				"Сотрудники",
+				"employees",
+				new ParametersFactory(UoW, (filters) =>
+				{
+					SelectableEntityParameter<Employee> resultAlias = null;
+					var query = UoW.Session.QueryOver<Employee>();
+
+					if(filters != null && filters.Any())
+					{
+						foreach(var f in filters)
+						{
+							query.Where(f());
+						}
+					}
+
+					var employeeProjection = CustomProjections.Concat_WS(
+						" ",
+						Projections.Property<Employee>(x => x.LastName),
+						Projections.Property<Employee>(x => x.Name),
+						Projections.Property<Employee>(x => x.Patronymic)
+					);
+
+					query.SelectList(list => list
+							.Select(x => x.Id).WithAlias(() => resultAlias.EntityId)
+							.Select(employeeProjection).WithAlias(() => resultAlias.EntityTitle))
+						.OrderBy(x => x.LastName).Asc
+						.OrderBy(x => x.Name).Asc
+						.OrderBy(x => x.Patronymic).Asc
+						.TransformUsing(Transformers.AliasToBean<SelectableEntityParameter<Employee>>());
+					
+					return query.List<SelectableParameter>();
+				})
+			);
+
+			SelectableParameterReportFilterViewModel = new SelectableParameterReportFilterViewModel(_filter);
 		}
 	}
 }
